recipe/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
psycopg2==2.8.5
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.0.1
